migtd: use feature `main` for conditional compilation of mods
Some of modules can only run in TDX environment and be built with
baremetal targets such as `x86_64-unknown-none`.

To make `migtd` crate usable by other tools/libs, split the definitions
and implementations in `ratls` and `migration` modules and use `main`
feauture to seperate them out.
