+++ .github/workflows/main.yml
@@ -63,7 +63,7 @@ jobs:
         run: cargo image
 
       - name: Build MigTD binary with a default production migration policy(virtio-serial)
-        run: cargo image --no-default-features --features stack-guard,virtio-serial
+        run: cargo image --no-default-features --features main,stack-guard,virtio-serial
 
   build_migtd_debug:
     name: Build final migtd - debug
@@ -110,4 +110,4 @@ jobs:
         run: cargo image --debug
 
       - name: Build MigTD binary with a default production migration policy(virtio-serial)
-        run: cargo image --debug --no-default-features --features stack-guard,virtio-serial
+        run: cargo image --debug --no-default-features --features main,stack-guard,virtio-serial
