+++ src/migtd/Cargo.toml
@@ -11,7 +11,7 @@ required-features = ["main"]
 [dependencies]
 bitfield = "0.13.2"
 anyhow = { version = "1.0.68", default-features = false }
-attestation = { path = "../attestation", default-features = false }
+attestation = { path = "../attestation", default-features = false, optional = true }
 cc-measurement = { path = "../../deps/td-shim/cc-measurement"}
 crypto = { path = "../crypto" }
 lazy_static = { version = "1.0", features = ["spin_no_std"] }
@@ -45,13 +45,13 @@ td-benchmark = { path = "../../deps/td-shim/devtools/td-benchmark", default-feat
 default = ["tdx"]
 cet-shstk = ["td-payload/cet-shstk"]
 coverage = ["minicov"]
-main = []
+main = ["tdx"]
 stack-guard = ["td-payload/stack-guard"]
 virtio-vsock = ["vsock/virtio-vsock"]
 virtio-serial = ["virtio_serial"]
 vmcall-interrupt = []
 vmcall-vsock = ["vsock/vmcall-vsock"]
-tdx = ["td-logger/tdx"]
+tdx = ["td-logger/tdx", "attestation"]
 test_heap_size = ["td-benchmark", "td-payload/test_heap_size"]
 test_stack_size = ["td-benchmark"]
 test_disable_ra_and_accept_all = ["attestation/test"] # Dangerous: can only be used for test purpose to bypass the remote attestation
